twitter_bot/requirements.txt
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.0
schedule==1.2.1
cryptocompare==0.7.6
